% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{aggr_duplicated_visits}
\alias{aggr_duplicated_visits}
\title{Aggregate duplicated visits.}
\usage{
aggr_duplicated_visits(df_visit)
}
\arguments{
\item{df_visit}{dataframe with columns: study_id, site_number, patnum, visit,
n_ae}
}
\value{
df_visit corrected
}
\description{
Internal function called by \link[=check_df_visit]{check_df_visit()}.
}
