% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{exp_implicit_missing_visits}
\alias{exp_implicit_missing_visits}
\title{Expose implicitly missing visits.}
\usage{
exp_implicit_missing_visits(df_visit, event_names = "ae")
}
\arguments{
\item{df_visit}{dataframe with columns: study_id, site_number, patnum, visit,
n_ae}

\item{event_names}{vector, contains the event names, default = "ae"}
}
\value{
df_visit corrected
}
\description{
Internal function called by \link[=check_df_visit]{check_df_visit()}.
}
