% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep_plot.R
\name{get_legend}
\alias{get_legend}
\title{replace cowplot::get_legend, to silence warning
Multiple components found; returning the first one. To return all, use `return_all = TRUE}
\usage{
get_legend(p)
}
\description{
replace cowplot::get_legend, to silence warning
Multiple components found; returning the first one. To return all, use `return_all = TRUE
}
\keyword{internal}
