% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{get_visit_med75}
\alias{get_visit_med75}
\title{get visit_med75}
\usage{
get_visit_med75(df_pat, method = "med75_adj")
}
\arguments{
\item{df_pat}{dataframe as returned by pat_aggr()}

\item{method}{character, one of c("med75", "med75_adj") defining method for
defining evaluation point visit_med75 (see details), Default: "med75_adj"}
}
\value{
dataframe
}
\description{
used by site_aggr()
}
