% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inframe.R
\name{p_adjust_bh_inframe}
\alias{p_adjust_bh_inframe}
\title{benjamini hochberg p value correction using table operations}
\usage{
p_adjust_bh_inframe(df_eval, col, suffix)
}
\description{
benjamini hochberg p value correction using table operations
}
\keyword{internal}
