% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{pat_aggr}
\alias{pat_aggr}
\title{aggregate visit to patient level}
\usage{
pat_aggr(df_visit)
}
\arguments{
\item{df_visit}{dataframe}
}
\value{
dataframe
}
\description{
used by site_aggr(), max_visit_per_pat
}
