% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_simaerep.R
\name{plot.simaerep}
\alias{plot.simaerep}
\title{plot AE under-reporting simulation results}
\usage{
\method{plot}{simaerep}(
  x,
  ...,
  study = NULL,
  what = "ur",
  n_sites = 16,
  df_visit = NULL,
  env = parent.frame()
)
}
\arguments{
\item{x}{simaerep object}

\item{...}{additional parameters passed to \link[=plot_study]{plot_study()} or
\link[=plot_visit_med75]{plot_visit_med75()}}

\item{study}{character specifying study to be plotted, Default: NULL}

\item{what}{one of c("ur", "med75"), specifying whether to plot site AE
under-reporting or visit_med75 values, Default: 'ur'}

\item{n_sites}{number of sites to plot, Default: 16}

\item{df_visit}{optional, pass original visit data if it cannot be retrieved
from parent environment, Default: NULL}

\item{env}{optional, pass environment from which to retrieve original visit
data, Default: parent.frame()}
}
\value{
ggplot object
}
\description{
generic plot function for simaerep objects
}
\details{
see \link[=plot_study]{plot_study()} and
\link[=plot_visit_med75]{plot_visit_med75()}
}
\examples{
\donttest{
df_visit <- sim_test_data_study(
  n_pat = 100,
  n_sites = 5,
  frac_site_with_ur = 0.4,
  ur_rate = 0.6
)

df_visit$study_id <- "A"

aerep <- simaerep(df_visit)

plot(aerep, what = "ur", study = "A")
plot(aerep, what = "med75", study = "A")
}
}
