% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_plot.R
\name{plot_study}
\alias{plot_study}
\title{plot ae development of study and sites, highlighting at risk sites}
\usage{
plot_study(df_visit, df_site, df_eval, study, df_al = NULL, n_sites = 16)
}
\arguments{
\item{df_visit}{dataframe}

\item{df_site}{dataframe}

\item{df_eval}{dataframe}

\item{study}{study}

\item{df_al}{dataframe containing study_roche, site_number, alert_level_site,
alert_level_study (optional), Default: NA}

\item{n_sites}{integer number of most at risk sites, Default: 16}
}
\value{
ggplot
}
\description{
" "
}
\details{
" "
}
\examples{
df_visit <- boot_sim_test_data_study(n_pat = 100, n_sites = 5,
    frac_site_with_ur = 0.4, ur_rate = 0.6)

df_visit$study_roche <- "A"
df_site <- site_aggr(df_visit)

df_sim_sites <- sim_sites(df_site, df_visit, r = 100)

df_eval <- eval_sites(df_sim_sites, r_sim_sites = 100)

plot_study(df_visit, df_site, df_eval, study = "A")
}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
