% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{prep_for_sim}
\alias{prep_for_sim}
\alias{sim_after_prep}
\title{prepare data for simulation}
\usage{
prep_for_sim(df_site, df_visit)

sim_after_prep(
  df_sim_prep,
  r = 1000,
  poisson_test = FALSE,
  prob_lower = TRUE,
  .progress = FALSE
)
}
\value{
dataframe

dataframe
}
\description{
collect AEs per patient at visit_med75 for site and study as a vector of integers

collect AEs per patient at visit_med75 for site and study as a vector of integers
}
\examples{
df_visit <- sim_test_data_study(
   n_pat = 100,
   n_sites = 5,
   frac_site_with_ur = 0.4,
   ur_rate = 0.2
)

df_visit$study_id <- "A"

df_site <- site_aggr(df_visit)

df_prep <- prep_for_sim(df_site, df_visit)
df_prep
df_visit <- sim_test_data_study(
   n_pat = 100,
   n_sites = 5,
   frac_site_with_ur = 0.4,
   ur_rate = 0.2
)

df_visit$study_id <- "A"

df_site <- site_aggr(df_visit)

df_prep <- prep_for_sim(df_site, df_visit)
df_prep
}
