% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_orivisit.R
\name{print.orivisit}
\alias{print.orivisit}
\title{Print method for orivisit objects}
\usage{
\method{print}{orivisit}(x, ..., n = 10)
}
\arguments{
\item{x}{An object of class 'orivisit'}

\item{...}{Additional arguments passed to print (not used)}

\item{n}{Number of rows to display from the data frame (default: 10)}
}
\description{
Print method for orivisit objects
}
\keyword{internal}
