% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_simaerep.R
\name{print.simaerep}
\alias{print.simaerep}
\title{Print method for simaerep objects}
\usage{
\method{print}{simaerep}(x, ..., n = 10)
}
\arguments{
\item{x}{An object of class 'simaerep'}

\item{...}{Additional arguments passed to print (not used)}

\item{n}{Number of rows to display from df_eval (default: 5)}
}
\description{
Print method for simaerep objects
}
