% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{prob_lower_site_ae_vs_study_ae}
\alias{prob_lower_site_ae_vs_study_ae}
\title{calculate probability for drawing a site sample with lower mean aes
from total study pool for vector with site AEs vs vector with study AEs}
\usage{
prob_lower_site_ae_vs_study_ae(site_ae, study_ae, r = 1000, parallel = F)
}
\arguments{
\item{site_ae}{vector with AE numbers}

\item{study_ae}{vector with AE numbers}

\item{r}{integer, denotes number of simulations, default = 1000}

\item{parallel}{logical, toggles parallel processing on and of, default = F}
}
\value{
pval
}
\description{
sets pvalue=1 if mean AE site is greater than mean AE study
}
\details{
" "
}
\examples{
prob_lower_site_ae_vs_study_ae(
  site_ae = c(5, 3, 3, 2, 1, 6),
  study_ae = c(9, 8, 7, 9, 6, 7, 8),
  parallel = FALSE
)
}
\seealso{
\code{\link[purrr]{safely}}
}
