% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{purrr_bar}
\alias{purrr_bar}
\title{execute a purrr or furrr function with a progress bar}
\usage{
purrr_bar(
  .purrr,
  ...,
  .f,
  .f_args = list(),
  .purrr_args = list(),
  .steps,
  .slow = FALSE,
  .progress = TRUE
)
}
\arguments{
\item{.purrr}{purrr or furrr function}

\item{...}{iterable arguments passed to .purrr}

\item{.f}{function to be executed over iterables}

\item{.f_args}{list of arguments passed to .f, Default: list()}

\item{.steps}{integer number of iterations}

\item{.slow}{logical slows down execution, Default: FALSE}

\item{.progress}{logical, show progress bar, Default: TRUE}

\item{.purr_args}{list of arguments passed to .purrr, Default: list()}
}
\description{
call still needs to be wrapped in with_progress()
}
\examples{

with_progress(purrr_bar(purrr::map, rep(0.25, 5), .f = Sys.sleep, .steps = 5))
with_progress(purrr_bar(purrr::walk, rep(0.25, 5), .f = Sys.sleep, .steps = 5))
with_progress(purrr_bar(purrr::walk, rep(0.25, 5), .f = Sys.sleep, .steps = 5,
              .progress = FALSE))

with_progress(
  purrr_bar(
    purrr::map2,
    rep(1, 5), rep(2, 5),
    .f = `+`,
    .steps = 5,
    .slow = TRUE
 )
)

}
