% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_simaerep.R
\name{remap_col_names}
\alias{remap_col_names}
\title{renames internal simaerep col_names to externally applied colnames}
\usage{
remap_col_names(df, col_names)
}
\description{
renames internal simaerep col_names to externally applied colnames
}
\keyword{internal}
