% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_test_data.R
\name{sim_pat}
\alias{sim_pat}
\title{simulate patients and events for sites
supports constant and non-constant event rates}
\usage{
sim_pat(
  vs_max,
  vs_sd,
  is_out,
  event_per_visit_mean,
  event_rates,
  event_names,
  factor_event_rate
)
}
\description{
simulate patients and events for sites
supports constant and non-constant event rates
}
\keyword{internal}
