% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_test_data.R
\name{sim_scenario}
\alias{sim_scenario}
\title{simulate single scenario}
\usage{
sim_scenario(n_ae_site, n_ae_study, frac_pat_with_ur, ur_rate)
}
\arguments{
\item{n_ae_site}{integer vector}

\item{n_ae_study}{integer vector}

\item{frac_pat_with_ur}{double}

\item{ur_rate}{double}
}
\value{
list
}
\description{
internal function called by simulate_scenarios()
}
\examples{
sim_scenario(c(5,5,5,5), c(8,8,8,8), 0.2, 0.5)
sim_scenario(c(5,5,5,5), c(8,8,8,8), 0.75, 0.5)
sim_scenario(c(5,5,5,5), c(8,8,8,8), 1, 0.5)
sim_scenario(c(5,5,5,5), c(8,8,8,8), 1, 1)
sim_scenario(c(5,5,5,5), c(8,8,8,8), 0, 0.5)
sim_scenario(c(5,5,5,5), c(8,8,8,8), 2, 0.5)
}
