% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_test_data.R
\name{sim_test_data_events}
\alias{sim_test_data_events}
\title{simulate test data events}
\usage{
sim_test_data_events(
  n_pat = 100,
  n_sites = 5,
  ae_per_visit_mean = 0.5,
  ae_rates = c(NULL),
  event_names = list("ae")
)
}
\arguments{
\item{n_pat}{integer, number of patients, Default: 100}

\item{n_sites}{integer, number of sites, Default: 5}

\item{ae_per_visit_mean}{mean event per visit per patient, Default: 0.5}

\item{ae_rates}{vector with visit-specific event rates, Default: Null}

\item{event_names}{vector, contains the event names, default = "ae"}
}
\value{
tibble with columns site_number, patnum, is_ur, max_visit_mean,
max_visit_sd, visit, and event data (events_per_visit_mean and n_events)
}
\description{
generates multi-event data using sim_test_data_study()
}
