% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_test_data.R
\name{sim_ur}
\alias{sim_ur}
\title{simulate under-reporting}
\usage{
sim_ur(df_visit, study_id, site_number, ur_rate)
}
\arguments{
\item{df_visit, }{dataframe}

\item{study_id, }{character}

\item{site_number, }{character}

\item{ur_rate, }{double}
}
\description{
we remove a fraction of AEs from a specific site
}
\details{
we determine the absolute number of AEs per patient to remove.
Then them remove them at the first visit.
We intentionally allow fractions
}
