% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inframe.R
\name{simaerep_inframe}
\alias{simaerep_inframe}
\title{simulate in dataframe}
\usage{
simaerep_inframe(
  df_visit,
  r = 1000,
  df_r = NULL,
  under_only = FALSE,
  method_visit_med75 = NULL
)
}
\arguments{
\item{df_visit}{data frame with columns: study_id, site_number, patnum, visit,
n_ae}

\item{df_r}{table with one column containing sequence of repetitions, provide
for in database mode}

\item{under_only}{logical compute under-reporting probabilities only,
superseeds under_only parameter passed to \link[=eval_sites]{eval_sites()} and
\link[=sim_sites]{sim_sites()}, Default: TRUE}
}
\description{
simulate in dataframe
}
\examples{
df_visit <- sim_test_data_study(
 n_pat = 100,
 n_sites = 5,
 frac_site_with_ur = 0.4,
 ur_rate = 0.6
)
df_visit$study_id <- "A"

simaerep_inframe(df_visit)
simaerep_inframe(df_visit, method_visit_med75 = "med75_adj")
\donttest{
# Database
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = ":memory:")
df_r <- tibble::tibble(rep = seq(1, 1000))

dplyr::copy_to(con, df_visit, "visit")
dplyr::copy_to(con, df_r, "r")

tbl_visit <- dplyr::tbl(con, "visit")
tbl_r <- dplyr::tbl(con, "r")

simaerep_inframe(tbl_visit, df_r = tbl_r)
simaerep_inframe(tbl_visit, df_r = tbl_r, method_visit_med75 = "med75_adj")

DBI::dbDisconnect(con)
}
}
\keyword{internal}
