% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{ttest_site_ae_vs_study_ae}
\alias{ttest_site_ae_vs_study_ae}
\title{ttest for vector with site AEs vs vector with study AEs}
\usage{
ttest_site_ae_vs_study_ae(site_ae, study_ae)
}
\arguments{
\item{site_ae}{vector with AE numbers}

\item{study_ae}{vector with AE numbers}
}
\value{
pval
}
\description{
sets pvalue=1 if mean AE site is greater than mean AE study or ttest gives error
}
\details{
" "
}
\examples{
ttest_site_ae_vs_study_ae(site_ae = c(5, 3, 3, 2, 1, 6), study_ae = c(9, 8, 7, 9, 6, 7, 8))
}
\seealso{
\code{\link[purrr]{safely}}
}
