% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{eval_sites}
\alias{eval_sites}
\title{evaluate sites}
\usage{
eval_sites(df_sim_sites, r_sim_sites = 1000)
}
\arguments{
\item{df_sim_sites}{dataframe generated by sim_sites()}

\item{r_sim_sites}{integer, simulation repeats for prob_lower for site
simulation, Default: 1000}
}
\value{
dataframe
}
\description{
evaluate site simulations by assigning a probability to each
scoring values using the score distributions of the study simulations. Then
calculate the P/fp ratio for each site.
}
\details{
zero probabilities are being replaced by the site minimum for
ecd_pval / 10 and by 1/r_sim_sites / 10
}
\examples{
df_visit <- boot_sim_test_data_study(n_pat = 100, n_sites = 5,
    frac_site_with_ur = 0.4, ur_rate = 0.6)

df_visit$study_id <- "A"
df_site <- site_aggr(df_visit)

df_sim_sites <- sim_sites(df_site, df_visit, r = 100)

df_eval <- eval_sites(df_sim_sites, r_sim_sites = 100)
df_eval
}
