% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{exp_implicit_missing_visits}
\alias{exp_implicit_missing_visits}
\title{expose implicitly missing visits}
\usage{
exp_implicit_missing_visits(df_visit)
}
\arguments{
\item{df_visit}{PARAM_DESCRIPTION}
}
\value{
df_visit corrected
}
\description{
called by check_df_visit()
}
