% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{get_site_mean_ae_dev}
\alias{get_site_mean_ae_dev}
\title{Get site mean ae development.}
\usage{
get_site_mean_ae_dev(df_visit, df_pat, df_site, event_names = c("ae"))
}
\arguments{
\item{df_visit}{dataframe}

\item{df_pat}{dataframe as returned by pat_aggr()}

\item{df_site}{dataframe as returned by site_aggr()}

\item{event_names}{vector, contains the event names, default = "ae"}
}
\value{
dataframe
}
\description{
Internal function used by \code{\link[=site_aggr]{site_aggr()}},
returns mean AE development from visit 0 to visit_med75.
}
\keyword{internal}
