% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{get_visit_med75}
\alias{get_visit_med75}
\title{Get visit_med75.}
\usage{
get_visit_med75(df_pat, method = "med75_adj", min_pat_pool = 0.2)
}
\arguments{
\item{df_pat}{dataframe as returned by \code{\link[=pat_aggr]{pat_aggr()}}}

\item{method}{character, one of c("med75", "med75_adj") defining method for
defining evaluation point visit_med75 (see details), Default: "med75_adj"}

\item{min_pat_pool}{double, minimum ratio of available patients available for
sampling. Determines maximum visit_med75 value see Details. Default: 0.2}
}
\value{
dataframe
}
\description{
Internal function used by \code{\link[=site_aggr]{site_aggr()}}.
}
