% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lint_package}
\alias{lint_package}
\title{Wrapper for \code{\link[lintr]{lint_package}}.}
\usage{
lint_package(path = ".", ...)
}
\arguments{
\item{path}{character}

\item{...}{additional arguments passed to \code{\link[lintr:lint]{lint()}}, e.g. \code{linters} or \code{cache}.}
}
\value{
" "
}
\description{
Throws error when code issues are detected and defines default
linting rules for this package.
}
\details{
lintr::line_length_linter(120), trailing_whitespace_linter = NULL,
cyclocomp_linter = lintr::cyclocomp_linter(25)
}
\seealso{
\code{\link[lintr]{lint_package}}
}
