% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{pat_aggr}
\alias{pat_aggr}
\title{Aggregate visit to patient level.}
\usage{
pat_aggr(df_visit)
}
\arguments{
\item{df_visit}{dataframe}
}
\value{
dataframe
}
\description{
Internal function used by \code{\link[=site_aggr]{site_aggr()}} and
\code{\link[=plot_visit_med75]{plot_visit_med75()}}, adds the maximum visit for each patient.
}
