% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inframe.R
\name{prune_to_visit_med75_inframe}
\alias{prune_to_visit_med75_inframe}
\title{prune visits to visit_med75 using table operations}
\usage{
prune_to_visit_med75_inframe(df_visit, df_site)
}
\arguments{
\item{df_visit}{Data frame with columns: study_id, site_number, patnum, visit,
n_ae.}

\item{df_site}{dataframe, as returned by \code{\link[=site_aggr]{site_aggr()}}}
}
\description{
prune visits to visit_med75 using table operations
}
\keyword{internal}
