% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{sim_sites}
\alias{sim_sites}
\title{calculate prob_lower and poisson.test pvalue for study sites}
\usage{
sim_sites(df_site, df_visit, r = 1000, poisson_test = TRUE, prob_lower = TRUE)
}
\arguments{
\item{df_site}{dataframe}

\item{df_visit}{dataframe}

\item{r}{integer, denotes number of simulations, default = 1000}

\item{poisson_test}{logical, calculates poisson.test pvalue}

\item{prob_lower}{logical, calculates probability for getting a lower value}
}
\value{
datafram with columns study_id, site_number, visit_med75, pval, prob_low
}
\description{
collects the number of AEs of all eligible patients from that
meet med75 criteria of site. Then uses calculates poisson.test pvalue and
probability of having a lower mean value using simulation.
}
\details{
" "
}
\examples{
df_visit <- boot_sim_test_data_study(n_pat = 100, n_sites = 5,
    frac_site_with_ur = 0.4, ur_rate = 0.2)
df_visit$study_id <- "A"
df_site <- site_aggr(df_visit)
df_sim_sites <- sim_sites(df_site, df_visit, r = 100)
}
