% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{sim_studies}
\alias{sim_studies}
\title{simulate studies}
\usage{
sim_studies(
  df_visit,
  df_site,
  r = r,
  ttest = TRUE,
  prob_lower = TRUE,
  r_prob_lower = 1000,
  parallel = FALSE,
  keep_ae = FALSE,
  min_n_pat_with_med75 = 1,
  studies = NULL,
  .progress = TRUE
)
}
\arguments{
\item{df_visit}{dataframe}

\item{df_site}{dataframe}

\item{r}{integer, denotes number of simulations, Default: 1000}

\item{ttest}{logical, calculates ttest pvalue, Default: T}

\item{prob_lower}{logical, calculates probability for getting a lower value,
Default: F}

\item{r_prob_lower}{nteger, denotes number of simulations for prob_lower
value calculation,, Default: 1000}

\item{parallel}{logical, toggles parallel processing on and of, Default: F}

\item{keep_ae}{logical, keep ae numbers in output dataframe memory increase
roughly 30 percent, Default: F}

\item{min_n_pat_with_med75}{integer, min number of patients with med75 at
site to simulate, Default: 1}

\item{studies}{vector with study names, Default: NULL}

\item{.progress}{logical, show progress bar}
}
\value{
dataframe
}
\description{
simulate pvalues and prob_lower values for testing
}
\details{
adds column with simulated probabilities for exqual or lower
mean_ae at visit_med75
}
\examples{
df_visit1 <- boot_sim_test_data_study(n_pat = 100, n_sites = 5,
                                      frac_site_with_ur = 0.4, ur_rate = 0.6)

df_visit1$study_roche <- "A"

df_visit2 <- boot_sim_test_data_study(n_pat = 1000, n_sites = 3,
                                      frac_site_with_ur = 0.2, ur_rate = 0.1)

df_visit2$study_roche <- "B"

df_visit <- bind_rows(df_visit1, df_visit2)

df_site <- site_aggr(df_visit)

sim_studies(ungroup(df_visit), ungroup(df_site), r = 3, keep_ae = TRUE)
}
