% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaerep.R
\name{site_aggr}
\alias{site_aggr}
\title{aggreagate from visit to site level}
\usage{
site_aggr(
  df_visit,
  col_study_id_str = "study_id",
  col_site_id_str = "site_number",
  col_patient_id_str = "patnum",
  col_n_ae_str = "n_ae",
  col_visit_str = "visit"
)
}
\arguments{
\item{df_visit}{dataframe}

\item{col_study_id_str}{column name for study id will be renamed to default
in output, Default: study_id}

\item{col_site_id_str}{column name for site id will be renamed to default in
output, Default: site_number}

\item{col_patient_id_str}{column name for patient id will be renamed to
default in output, Default: patnum}

\item{col_n_ae_str}{column name for number of ae will be renamed to default
in output, Default: n_ae}

\item{col_visit_str}{column name for consecutive visit number will be renamed
to default in output, Default: visit}
}
\value{
dataframe with the following columns:
\describe{
\item{\strong{study_id}}{study identification}
\item{\strong{site_number}}{site identification}
\item{\strong{n_patients}}{number of patients, site level}
\item{\strong{visit_med75}}{median(max(visit)) * 0.75}
\item{\strong{n_pat_with_med75}}{number of patients that meet visit_med75 criterion, site level}
\item{\strong{mean_ae_site_med75}}{mean AE at visit_med75, site level}
}
}
\description{
calculates visit_med75, n_pat_with_med75 and mean_ae_site_med75
}
\examples{
df_visit <- sim_test_data_study(n_pat = 100, n_sites = 5,
    frac_site_with_ur = 0.4, ur_rate = 0.6)
df_visit$study_id <- "A"
df_site <- site_aggr(df_visit)
df_site
}
